searchvarの返り値変更

+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: takanoraika <takanoraika@student.42.fr>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/14 11:20:30 by shima             #+#    #+#             */
-/*   Updated: 2022/10/09 10:42:41 by takanoraika      ###   ########.fr       */
+/*   Updated: 2022/10/11 13:18:10 by takanoraika      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ void	signal_set(void);
 void	ft_export(char **args);
 int		ft_echo(char **args);
 int		ft_isspace(int c);
-int		search_var(char *name);
+ssize_t	search_var(char *name);
 char	*return_name(char *var);
 
 void	error_exit(const char *s);
